Port java part to new non-uuid based rust core

+++ core/retro-bridge/src/util/display.rs
@@ -12,7 +12,7 @@ pub struct NativeDisplay {
 }
 
 impl NativeDisplay {
-    pub fn new(width: i32, height: i32, codec: i32, data_ptr: i32) -> Self {
+    pub fn new(width: i32, height: i32, codec: i32, data_ptr: i64) -> Self {
         let size = (width * height * 3) as usize;
         NativeDisplay {
             data: unsafe {
